requirements.txt
flask
pymongo
mysql-connector-python
